:sparkles: Add JSON Schema 2020-12 tool schemas and trace context to MCP (#221)
Lift MCP tool schemas to full JSON Schema 2020-12 (SEP-2106): derive a tool
outputSchema from the handler return annotation, return structuredContent on
tool calls, and emit self-contained input/output schemas that inline every
nested model under $defs with local #/$defs references and no external
dereferencing. Propagate W3C Trace Context (traceparent/tracestate/baggage)
through _meta and expose it as an injectable MCPTraceContext component
(SEP-414).

Refactor the schema layer to support this: extract a generic SchemaRegistry
with a bundle() helper for self-contained documents, move the OpenAPI-specific
resolution into an OpenAPISchemaRegistry, and merge the OpenAPI SchemaGenerator
into schemas/openapi.py (dropping schemas/generator.py).
