+++ flama/schemas/openapi.py
@@ -1,7 +1,15 @@
 import dataclasses
+import inspect
+import itertools
+import logging
 import typing as t
+from collections import defaultdict
 
-from flama import types
+import yaml
+
+from flama import routing, schemas, types, url
+from flama.schemas import data_structures as ds
+from flama.schemas.registry import SchemaInfo, SchemaRegistry
 
 __all__ = [
     "Schema",
@@ -30,9 +38,12 @@ __all__ = [
     "Components",
     "OpenAPI",
     "OpenAPISpec",
+    "OpenAPISchemaRegistry",
+    "SchemaGenerator",
 ]
 
 empty = types.Empty()
+logger = logging.getLogger(__name__)
 
 
 class Schema(types.JSONSchema): ...
@@ -408,3 +419,423 @@ class OpenAPISpec:
             return self.to_dict(dataclasses.asdict(obj))
 
         return obj
+
+
+class OpenAPISchemaRegistry(SchemaRegistry):
+    """Schema registry that resolves references against an OpenAPI ``#/components/schemas`` section."""
+
+    def _get_schema_references_from_schema(self, schema: "Schema | Reference") -> list[str]:
+        if isinstance(schema, Reference):
+            return [schema.ref]
+
+        result = []
+
+        if "$ref" in schema:
+            result.append(schema["$ref"])
+
+        if schema.get("type", "") == "array":
+            items = schema.get("items", {})
+            if isinstance(items, dict) and "$ref" in items:
+                result.append(items["$ref"])
+
+        for composer in ("allOf", "anyOf", "oneOf"):
+            composer_schemas = schema.get(composer, [])
+            if isinstance(composer_schemas, list):
+                result += [
+                    ref
+                    for x in composer_schemas
+                    if x and isinstance(x, dict)
+                    for ref in self._get_schema_references_from_schema(Schema(x))
+                ]
+
+        props = schema.get("properties", {})
+        if isinstance(props, dict):
+            result += [
+                ref
+                for x in props.values()
+                if x and isinstance(x, dict)
+                for ref in self._get_schema_references_from_schema(Schema(x))
+            ]
+
+        return result
+
+    def _get_schema_references_from_path(self, path: "Path") -> list[str]:
+        return [y for x in path.operations.values() for y in self._get_schema_references_from_operation(x)]
+
+    def _get_schema_references_from_operation(self, operation: "Operation") -> list[str]:
+        return [
+            *self._get_schema_references_from_operation_parameters(operation.parameters),
+            *self._get_schema_references_from_operation_request_body(operation.requestBody),
+            *self._get_schema_references_from_operation_callbacks(operation.callbacks),
+            *self._get_schema_references_from_operation_responses(operation.responses),
+        ]
+
+    def _get_schema_references_from_operation_responses(self, responses: "Responses") -> list[str]:
+        refs = []
+
+        for response in responses.values():
+            content = response.content
+            if not content:
+                continue
+            for media_type in content.values():
+                if not isinstance(media_type, MediaType):
+                    continue
+                schema = media_type.schema
+                if schema is None:
+                    continue
+                refs += self._get_schema_references_from_schema(schema)
+
+        return refs
+
+    def _get_schema_references_from_operation_callbacks(
+        self, callbacks: "dict[str, Callback | Reference] | None"
+    ) -> list[str]:
+        refs = []
+
+        if callbacks:
+            for callback in callbacks.values():
+                if isinstance(callback, Reference):
+                    refs.append(callback.ref)
+                else:
+                    for callback_path in callback.values():
+                        refs += self._get_schema_references_from_path(callback_path)
+
+        return refs
+
+    def _get_schema_references_from_operation_request_body(
+        self, request_body: "RequestBody | Reference | None"
+    ) -> list[str]:
+        refs = []
+
+        if request_body:
+            if isinstance(request_body, Reference):
+                refs.append(request_body.ref)
+            else:
+                for media_type in request_body.content.values():
+                    if not isinstance(media_type, MediaType):
+                        continue
+                    schema = media_type.schema
+                    if schema is None:
+                        continue
+                    refs += self._get_schema_references_from_schema(schema)
+
+        return refs
+
+    def _get_schema_references_from_operation_parameters(
+        self, parameters: "list[Parameter | Reference] | None"
+    ) -> list[str]:
+        refs = []
+
+        if parameters:
+            for param in parameters:
+                if isinstance(param, Reference):
+                    refs.append(param.ref)
+                elif param.schema is not None:
+                    refs += self._get_schema_references_from_schema(param.schema)
+
+        return refs
+
+    def used(self, spec: "OpenAPISpec") -> dict[int, SchemaInfo]:
+        """Generate a dict containing used schemas.
+
+        :param spec: API Schema.
+        :return: Used schemas.
+        """
+        # Get schemas from references in path
+        refs = {
+            x.split("/")[-1] for path in spec.spec.paths.values() for x in self._get_schema_references_from_path(path)
+        }
+        used_schemas = {id_: schema for id_, schema in self.items() if schema.name in refs}
+
+        # Get schemas from references in schema
+        partial_schemas = set(used_schemas.values())
+        refs = set([])
+        while partial_schemas:
+            schema = partial_schemas.pop()
+            refs_from_schema = {
+                ref.split("/")[-1]
+                for ref in self._get_schema_references_from_schema(Schema(schema.json_schema(self.names)))
+            }
+            partial_schemas |= {schema for schema in self.values() if schema.name in refs_from_schema}
+            refs |= refs_from_schema
+        used_schemas |= {id_: schema for id_, schema in self.items() if schema.name in refs}
+
+        # Get schemas from schema children
+        partial_schemas = set(used_schemas.values())
+        while partial_schemas:
+            for schema in (
+                ds.Schema(s).unique_schema for s in ds.Schema(partial_schemas.pop().schema).nested_schemas()
+            ):
+                partial_schemas.add(self[schema])
+                used_schemas[id(schema)] = self[schema]
+
+        return used_schemas
+
+    def get_openapi_ref(self, element: "ds.Schema", multiple: bool | None = None) -> "Schema | Reference":
+        """Builds the reference for a single schema or the array schema containing the reference.
+
+        :param element: Schema to use as reference.
+        :param multiple: True for building a schema containing an array of references instead of a single reference.
+        :return: Reference or array schema.
+        """
+        reference = f"#/components/schemas/{self[element].name}"
+
+        if multiple is True:
+            return Schema({"items": {"$ref": reference}, "type": "array"})
+        else:
+            return Reference(ref=reference)
+
+
+@dataclasses.dataclass(frozen=True)
+class EndpointInfo:
+    path: str
+    method: str
+    func: t.Callable = dataclasses.field(repr=False)
+    query_parameters: dict[str, ds.Parameter] = dataclasses.field(repr=False)
+    path_parameters: dict[str, ds.Parameter] = dataclasses.field(repr=False)
+    body_parameter: ds.Parameter | None = dataclasses.field(repr=False)
+    response_parameter: ds.Parameter = dataclasses.field(repr=False)
+
+
+class SchemaGenerator:
+    def __init__(self, spec: types.OpenAPISpec, schemas: dict[str, ds.Schema] | None = None):
+        self.spec = OpenAPISpec.from_spec(spec)
+
+        # Builtin definitions
+        self.schemas = OpenAPISchemaRegistry(schemas=schemas)
+
+    def get_endpoints(  # type: ignore[override]
+        self, routes: list[routing.BaseRoute], base_path: str = ""
+    ) -> dict[str, list[EndpointInfo]]:
+        """
+        Given the routes, yields the following information:
+
+        - path
+            eg: /users/
+        - http_method
+            one of 'get', 'post', 'put', 'patch', 'delete', 'options'
+        - func
+            method ready to extract the docstring
+
+        :param routes: A list of routes.
+        :param base_path: The base endpoints path.
+        :return: Data structure that contains metadata from every route.
+        """
+        endpoints_info: dict[str, list[EndpointInfo]] = defaultdict(list)
+
+        for route in routes:
+            path = str(url.Path(base_path) / route.path)
+
+            if isinstance(route, routing.Route) and route.include_in_schema:
+                if inspect.isfunction(route.endpoint) or inspect.ismethod(route.endpoint):
+                    for method in route.methods or ["GET"]:
+                        if method == "HEAD":
+                            continue
+
+                        endpoints_info[path].append(
+                            EndpointInfo(
+                                path=path,
+                                method=method.lower(),
+                                func=route.endpoint,
+                                query_parameters=route.parameters.query.get(method, {}),
+                                path_parameters=route.parameters.path.get(method, {}),
+                                body_parameter=route.parameters.body.get(method),
+                                response_parameter=route.parameters.response[method],
+                            )
+                        )
+                else:
+                    for method in [x.lower() for x in route.methods]:
+                        if not hasattr(route.endpoint, method):
+                            continue
+
+                        func = getattr(route.endpoint, method)
+                        endpoints_info[path].append(
+                            EndpointInfo(
+                                path=path,
+                                method=method.lower(),
+                                func=func,
+                                query_parameters=route.parameters.query.get(method.upper(), {}),
+                                path_parameters=route.parameters.path.get(method.upper(), {}),
+                                body_parameter=route.parameters.body.get(method.upper()),
+                                response_parameter=route.parameters.response[method.upper()],
+                            )
+                        )
+            elif isinstance(route, routing.Mount):
+                endpoints_info.update(self.get_endpoints(route.routes, base_path=path))
+
+        return endpoints_info
+
+    def _build_endpoint_parameters(
+        self, endpoint: EndpointInfo, metadata: dict[str, t.Any]
+    ) -> list[Parameter | Reference] | None:
+        if not endpoint.query_parameters and not endpoint.path_parameters:
+            return None
+
+        return [
+            Parameter(
+                schema=Schema(field.field.json_schema),
+                name=field.name,
+                in_=field.location.name,
+                required=field.required,
+                **{
+                    x: y.get(x)
+                    for y in [x for x in metadata.get("parameters", []) if x.get("name") == field.name]
+                    if y
+                    for x in (
+                        "description",
+                        "deprecated",
+                        "allowEmptyValue",
+                        "style",
+                        "explode",
+                        "allowReserved",
+                        "example",
+                    )
+                },
+            )
+            for field in itertools.chain(endpoint.query_parameters.values(), endpoint.path_parameters.values())
+        ]
+
+    def _build_endpoint_body(self, endpoint: EndpointInfo, metadata: dict[str, t.Any]) -> RequestBody | None:
+        content = {k: v for k, v in metadata.get("requestBody", {}).get("content", {}).items()}
+
+        if endpoint.body_parameter:
+            if endpoint.body_parameter.schema.schema not in self.schemas:
+                self.schemas.register(schema=endpoint.body_parameter.schema.schema)
+
+            content["application/json"] = MediaType(
+                schema=self.schemas.get_openapi_ref(
+                    endpoint.body_parameter.schema.schema, multiple=endpoint.body_parameter.multiple
+                ),
+            )
+
+        if not content:
+            return None
+
+        return RequestBody(
+            content=content,
+            **{x: metadata.get("requestBody", {}).get(x) for x in ("description", "required")},
+        )
+
+    def _build_endpoint_responses(self, endpoint: EndpointInfo, metadata: dict[str, t.Any]) -> Responses:
+        responses = metadata.get("responses", {})
+        try:
+            main_response_code = next(iter(responses.keys()))
+            assert 100 <= int(main_response_code) < 600
+        except (ValueError, AssertionError, StopIteration):
+            main_response_code = 200
+            responses[main_response_code] = {
+                "description": "Description not provided.",
+            }
+            logger.warning(
+                'OpenAPI description not provided in docstring for main response in endpoint "%s", adding a standard '
+                '"%s" response',
+                main_response_code,
+                endpoint.path,
+            )
+
+        if endpoint.response_parameter.schema.schema:
+            if endpoint.response_parameter.schema.schema not in self.schemas:
+                self.schemas.register(schema=endpoint.response_parameter.schema.schema)
+
+            responses[main_response_code]["content"] = {
+                **responses[main_response_code].get("content", {}),
+                "application/json": {
+                    "schema": self.schemas.get_openapi_ref(
+                        endpoint.response_parameter.schema.schema, multiple=endpoint.response_parameter.multiple
+                    )
+                },
+            }
+
+        responses["default"] = {
+            "description": "Unexpected error.",
+            **responses.get("default", {}),
+            "content": {
+                "application/json": {
+                    "schema": self.schemas.get_openapi_ref(schemas.schemas.core.APIError, multiple=False)
+                }
+            },
+        }
+
+        return Responses(
+            {
+                str(code): Response(
+                    description=response["description"],
+                    headers=response.get("headers"),
+                    content={
+                        mime: MediaType(
+                            schema=media_type.get("schema"),
+                            example=media_type.get("example"),
+                            examples=media_type.get("examples"),
+                            encoding=media_type.get("encoding"),
+                        )
+                        for mime, media_type in response.get("content", {}).items()
+                    },
+                    links=response.get("links"),
+                )
+                for code, response in responses.items()
+            }
+        )
+
+    def _parse_docstring(self, func: t.Callable) -> dict[t.Any, t.Any]:
+        """Given a function, parse the docstring as YAML and return a dictionary of info.
+
+        :param func: Function to analyze docstring.
+        :return: Schema extracted.
+        """
+        try:
+            # It's possible to define a standard docstring along with the schema definition, for doing so the schema
+            # should start with a line with three dashes "---" as it's the usual notation for starting a yaml file.
+            schema = yaml.safe_load((func.__doc__ or "").split("---")[-1])
+        except AttributeError:
+            schema = None
+
+        if not isinstance(schema, dict):
+            raise ValueError
+
+        return schema
+
+    def get_operation_schema(self, endpoint: EndpointInfo) -> Operation:
+        try:
+            docstring_info = self._parse_docstring(endpoint.func)
+        except ValueError:
+            docstring_info = {}
+
+        # Query and Path parameters
+        parameters = self._build_endpoint_parameters(endpoint, docstring_info)
+
+        # Body
+        request_body = self._build_endpoint_body(endpoint, docstring_info)
+
+        # Response
+        responses = self._build_endpoint_responses(endpoint, docstring_info)
+
+        return Operation(
+            responses=responses,
+            parameters=parameters,
+            requestBody=request_body,
+            **{
+                x: docstring_info.get(x)
+                for x in ("tags", "summary", "description", "externalDocs", "operationId", "deprecated")
+            },
+        )
+
+    def get_api_schema(self, routes: list[routing.BaseRoute]) -> dict[str, t.Any]:
+        endpoints_info = self.get_endpoints(routes)
+
+        for path, endpoints in endpoints_info.items():
+            operations = {}
+            for endpoint in endpoints:
+                try:
+                    operations[endpoint.method] = self.get_operation_schema(endpoint)
+                except Exception:
+                    logger.error("Cannot generate schema for endpoint %s", endpoint)
+
+            if operations:
+                self.spec.add_path(path, Path(**operations))  # type: ignore[arg-type]
+
+        for schema in self.schemas.used(self.spec).values():
+            self.spec.add_schema(schema.name, Schema(schema.json_schema(self.schemas.names)))
+
+        api_schema: dict[str, t.Any] = self.spec.to_dict()
+
+        return api_schema
